fix order of jnz operands
