+++ tests/functional/venom/parser/test_parsing.py
@@ -1,48 +1,8 @@
+from tests.venom_utils import assert_ctx_eq
 from vyper.venom.basicblock import IRBasicBlock, IRInstruction, IRLabel, IRLiteral, IRVariable
 from vyper.venom.context import IRContext
 from vyper.venom.function import IRFunction
 from vyper.venom.parser import parse_venom
-
-# TODO: Refactor tests with these helpers
-
-
-def instructions_eq(i1: IRInstruction, i2: IRInstruction) -> bool:
-    return i1.output == i2.output and i1.opcode == i2.opcode and i1.operands == i2.operands
-
-
-def assert_bb_eq(bb1: IRBasicBlock, bb2: IRBasicBlock):
-    assert bb1.label.value == bb2.label.value
-    assert len(bb1.instructions) == len(bb2.instructions)
-    for i1, i2 in zip(bb1.instructions, bb2.instructions):
-        assert instructions_eq(i1, i2), f"[{i1}] != [{i2}]"
-
-
-def assert_fn_eq(fn1: IRFunction, fn2: IRFunction):
-    assert fn1.name.value == fn2.name.value
-    assert fn1.last_variable == fn2.last_variable
-    assert len(fn1._basic_block_dict) == len(fn2._basic_block_dict)
-
-    for name1, bb1 in fn1._basic_block_dict.items():
-        assert name1 in fn2._basic_block_dict
-        assert_bb_eq(bb1, fn2._basic_block_dict[name1])
-
-    # check function entry is the same
-    assert fn1.entry.label == fn2.entry.label
-
-
-def assert_ctx_eq(ctx1: IRContext, ctx2: IRContext):
-    assert ctx1.last_label == ctx2.last_label
-    assert len(ctx1.functions) == len(ctx2.functions)
-    for label1, fn1 in ctx1.functions.items():
-        assert label1 in ctx2.functions
-        assert_fn_eq(fn1, ctx2.functions[label1])
-
-    # check entry function is the same
-    assert next(iter(ctx1.functions.keys())) == next(iter(ctx2.functions.keys()))
-
-    assert len(ctx1.data_segment) == len(ctx2.data_segment)
-    for d1, d2 in zip(ctx1.data_segment, ctx2.data_segment):
-        assert instructions_eq(d1, d2), f"data: [{d1}] != [{d2}]"
 
 
 def test_single_bb():
@@ -70,7 +30,7 @@ def test_multi_bb_single_fn():
     function start {
         start:
             %1 = callvalue
-            jnz @fine, @has_callvalue, %1
+            jnz %1, @fine, @has_callvalue
         fine:
             %2 = calldataload 4
             %4 = add %2, 279387
@@ -89,7 +49,7 @@ def test_multi_bb_single_fn():
 
     start_bb = start_fn.get_basic_block("start")
     start_bb.append_instruction("callvalue", ret=IRVariable("1"))
-    start_bb.append_instruction("jnz", IRVariable("1"), IRLabel("has_callvalue"), IRLabel("fine"))
+    start_bb.append_instruction("jnz", IRVariable("1"), IRLabel("fine"), IRLabel("has_callvalue"))
 
     start_fn.append_basic_block(fine_bb := IRBasicBlock(IRLabel("fine"), start_fn))
     fine_bb.append_instruction("calldataload", IRLiteral(4), ret=IRVariable("2"))
@@ -160,7 +120,7 @@ def test_multi_function():
         check_cv:
             %1 = callvalue
             %2 = param
-            jnz @no_value, @has_value, %1
+            jnz %1, @has_value, @no_value
         no_value:
             ret %2
         has_value:
@@ -218,7 +178,7 @@ def test_multi_function_and_data():
         check_cv:
             %1 = callvalue
             %2 = param
-            jnz @no_value, @has_value, %1
+            jnz %1, @has_value, @no_value
         no_value:
             ret %2
         has_value:
